# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "实验室人脸识别打卡系统",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/face/enrollment-sessions": {
      "post": {
        "tags": [
          "人脸档案"
        ],
        "summary": "Create Enrollment",
        "operationId": "create_enrollment_api_v1_face_enrollment_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/face/enrollment-sessions/{enrollment_id}/frames": {
      "post": {
        "tags": [
          "人脸档案"
        ],
        "summary": "Add Enrollment Frames",
        "operationId": "add_enrollment_frames_api_v1_face_enrollment_sessions__enrollment_id__frames_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "enrollment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Enrollment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_add_enrollment_frames_api_v1_face_enrollment_sessions__enrollment_id__frames_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameProcessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/face/enrollment-sessions/{enrollment_id}/submit": {
      "post": {
        "tags": [
          "人脸档案"
        ],
        "summary": "Submit Enrollment",
        "operationId": "submit_enrollment_api_v1_face_enrollment_sessions__enrollment_id__submit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "enrollment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Enrollment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaceProfilePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/face-profile": {
      "get": {
        "tags": [
          "人脸档案"
        ],
        "summary": "My Face Profiles",
        "operationId": "my_face_profiles_api_v1_me_face_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FaceProfilePublic"
                  },
                  "type": "array",
                  "title": "Response My Face Profiles Api V1 Me Face Profile Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/face-profiles": {
      "get": {
        "tags": [
          "人脸档案"
        ],
        "summary": "Admin Face Profiles",
        "operationId": "admin_face_profiles_api_v1_admin_face_profiles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FaceProfileStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FaceProfilePublic_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/face-profiles/{profile_id}/approve": {
      "post": {
        "tags": [
          "人脸档案"
        ],
        "summary": "Approve Profile",
        "operationId": "approve_profile_api_v1_admin_face_profiles__profile_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaceProfilePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/face-profiles/{profile_id}/reject": {
      "post": {
        "tags": [
          "人脸档案"
        ],
        "summary": "Reject Profile",
        "operationId": "reject_profile_api_v1_admin_face_profiles__profile_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectFaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FaceProfilePublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/face-profiles/{profile_id}/revoke": {
      "post": {
        "tags": [
          "人脸档案"
        ],
        "summary": "Revoke Profile",
        "operationId": "revoke_profile_api_v1_admin_face_profiles__profile_id__revoke_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_add_enrollment_frames_api_v1_face_enrollment_sessions__enrollment_id__frames_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_add_enrollment_frames_api_v1_face_enrollment_sessions__enrollment_id__frames_post"
      },
...
      "EnrollmentCreate": {
        "properties": {
          "target_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target User Id"
          },
          "mode": {
            "$ref": "#/components/schemas/EnrollmentMode",
            "default": "SELF"
          }
        },
        "type": "object",
        "title": "EnrollmentCreate"
      },
      "EnrollmentMode": {
        "type": "string",
        "enum": [
          "SELF",
          "ADMIN"
        ],
        "title": "EnrollmentMode"
      },
      "EnrollmentPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "profile_id": {
            "type": "string",
            "format": "uuid",
            "title": "Profile Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "status": {
            "$ref": "#/components/schemas/EnrollmentStatus"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "profile_id",
          "user_id",
          "status",
          "expires_at",
          "created_at"
        ],
        "title": "EnrollmentPublic"
      },
      "EnrollmentStatus": {
        "type": "string",
        "enum": [
          "COLLECTING",
          "SUBMITTED",
          "COMPLETED",
          "EXPIRED"
        ],
        "title": "EnrollmentStatus"
      },
...
      "FaceProfilePublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "status": {
            "$ref": "#/components/schemas/FaceProfileStatus"
          },
          "mode": {
            "$ref": "#/components/schemas/EnrollmentMode"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "liveness_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Liveness Score"
          },
          "live_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live Verified At"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "template_count": {
            "type": "integer",
            "title": "Template Count",
            "default": 0
          },
          "user_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Name"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "status",
          "mode",
          "model_version",
          "quality_score",
          "liveness_score",
          "live_verified_at",
          "rejection_reason",
          "created_at",
          "submitted_at",
          "reviewed_at"
        ],
        "title": "FaceProfilePublic"
      },
      "FaceProfileStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "PENDING",
          "ACTIVE",
          "REJECTED",
          "REPLACED",
          "REVOKED"
        ],
        "title": "FaceProfileStatus"
      },
      "FrameProcessResponse": {
        "properties": {
          "accepted": {
            "type": "integer",
            "title": "Accepted"
          },
          "rejected": {
            "type": "integer",
            "title": "Rejected"
          },
          "template_count": {
            "type": "integer",
            "title": "Template Count"
          },
          "average_quality": {
            "type": "number",
            "title": "Average Quality"
          },
          "liveness_score": {
            "type": "number",
            "title": "Liveness Score"
          },
          "messages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "accepted",
          "rejected",
          "template_count",
          "average_quality",
          "liveness_score",
          "messages"
        ],
        "title": "FrameProcessResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
...
      "Page_FaceProfilePublic_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FaceProfilePublic"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[FaceProfilePublic]"
      },
...
      "RejectFaceRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 2,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "RejectFaceRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}